Created go.mod file and used different files for main function- main.go and hash function -  hash.go also added comments above each function
