+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/brevo-assignments/Backend/Hash-assignment
+
+go 1.24.1
